Move StyledText & its Ops class to model package
